Minor fixes related to build executables (#7)
On MacOS, the plugin couldn't find the msbuild executable even though it was on path. This was fixed by scanning for msbuild instead of MSBuild.exe on unix OSes.

Also on MacOS, sonar scanner execution would fail as it couldn't access the inner bin/sonarscanner script. This was fixed by giving execution permission to every file inside the .../bin folder in the sonarscanner installation folder.

Finally, now it is possible to supply extra arguments to both dotnet and msbuild executables using the extraArgs property in the BuildSolution task.

Co-authored-by: Joaquim Neto <joaquim.neto@wooga.net>